requirements.txt
kagglehub
python-dotenv
scikit-learn
numpy
torch
torchvision
matplotlib
seaborn
fastapi
uvicorn[standard]
python-multipart
redis
slowapi
pydantic
